Replaced the Makefile with a CMakefile that looks much cleaner
And yet, the CMakefile might be harder to customise.

That said, I'll try to add some targets like Windows systems, Tizen
systems and, maybe, Firefox OS systems in the future.
These kind of targets cannot be added nicely with simple Makefile,
not withstanding that CMake tends to integrate much better with IDE,
be it Qt Creator or Visual Studio.

Signed-off-by: Myy <myy@miouyouyou.fr>