It's working !
But... The API is fugly...

There's still a lot of mess to deal with...
For the matter, the fonts still to be "pre-packed" instead of being
generated on-the-fly, which would alleviate the whole "textures" and
"pack files" retardations here and there.

Same thing for the cursor textures, which should be imported from
ICO or BMP files directly, instead of myyraw bullshit, which only
mattered for my own projects, at the time.

Also, I started to add the SDF code there... I need to find a good
SDF atlas generator, integrate it to the project and use it instead
of the current generator (or in parallel with the current generator).

And, at last, the "text atlases", implementation details and text
area definitions should be "encapsulated" behind major text areas
collections objects, which should draw everything at once, and help
create and manipulate text areas through simple "ID" handles, instead
of having to deal with structures and what nots.

Anyway, it's working... but it still requires a lot of work...
I'll retry to rework the API, once I get a little less busy, in order
to at least provide a nice solution for the next person asking
"How to display text in OpenGL ES" ( #1 ).

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>